Clean up logs and tests
